update dev, change query, query params

+++ controllers/RoleController.js
@@ -21,6 +21,7 @@ export const createRole = async (req, res) => {
             const checkRole = await Role.findOne({ where: { role_name } })
             if (checkRole) return payload(400, false, "Role already exist", null, res)
         }
+
         const role = await Role.create({ role_name })
         return payload(200, true, "Role created", role, res)
     } catch (error) {
@@ -32,8 +33,13 @@ export const updateRole = async (req, res) => {
     try {
         const { id } = req.params
         const { role_name } = req.body
+
+        const checkRole = await Role.findOne({ where: { role_name } })
+        if (checkRole) return payload(400, false, "Role already exist", null, res)
+
         const role = await Role.update({ role_name }, { where: { id } })
         if (!role) return payload(400, false, "Role not found", null, res)
+
         return payload(200, true, "Role updated", null, res)
     } catch (error) {
         return payload(500, false, error.message, null, res)
@@ -43,8 +49,13 @@ export const updateRole = async (req, res) => {
 export const deleteRole = async (req, res) => {
     try {
         const { id } = req.params
+
+        const checkRole = await Role.findOne({ where: { id } })
+        if (!checkRole) return payload(400, false, "Role not found", null, res)
+
         const role = await Role.destroy({ where: { id } })
         if (role === 0) return payload(400, false, "Role not found", null, res)
+
         return payload(200, true, "Role deleted", null, res)
     } catch (error) {
         return payload(500, false, error.message, null, res)
