+++ controllers/SatuanKerjaController.js
@@ -1,5 +1,6 @@
 import payload from "../response_format.js";
 import SatuanKerja from "../models/SatuanKerjaModel.js";
+import { or } from "sequelize";
 
 export const getAllSatuanKerja = async (req, res) => {
     try {
@@ -20,14 +21,21 @@ export const getAllSatuanKerja = async (req, res) => {
 export const createSatuanKerja = async (req, res) => {
     try {
         const { kode_satuan_kerja, nama_satuan_kerja, email_satuan_kerja, latitude,
-            longitude,alamat_web_satuan_kerja, alamat_satuan_kerja, no_telp_satuan_kerja, no_fax_satuan_kerja, kode_pos_satuan_kerja, kode_prov, kode_kab, kode_kec, kode_desa } = req.body
+            longitude, alamat_web_satuan_kerja, alamat_satuan_kerja, no_telp_satuan_kerja, no_fax_satuan_kerja, kode_pos_satuan_kerja, kode_prov, kode_kab, kode_kec, kode_desa } = req.body
 
         const checkSatuanKerja = await SatuanKerja.findOne({
             where: {
                 nama_satuan_kerja: nama_satuan_kerja
             }
         })
-        if (checkSatuanKerja)  return payload(400, false, "Satuan kerja sudah terdaftar", null, res)
+        if (checkSatuanKerja) return payload(400, false, "Satuan kerja sudah terdaftar", null, res)
+
+        const checkKodeSatuanKerja = await SatuanKerja.findOne({
+            where: {
+                kode_satuan_kerja: kode_satuan_kerja
+            }
+        })
+        if (checkKodeSatuanKerja) return payload(400, false, "Kode satuan kerja sudah terdaftar", null, res)
 
         const satuanKerja = await SatuanKerja.create({
             kode_satuan_kerja,
@@ -75,11 +83,11 @@ export const updateSatuanKerja = async (req, res) => {
         const checkSatuanKerja = await SatuanKerja.findOne({
             where: {
                 nama_satuan_kerja: nama_satuan_kerja
-            }
+            },
         })
-        if (checkSatuanKerja)  return payload(400, false, "Satuan kerja sudah terdaftar", null, res)
+        if (checkSatuanKerja) return payload(400, false, "Satuan kerja sudah terdaftar", null, res)
 
-        const satuanKerja = await SatuanKerja.update({
+        await SatuanKerja.update({
             kode_satuan_kerja,
             nama_satuan_kerja,
             latitude,
@@ -108,11 +116,20 @@ export const updateSatuanKerja = async (req, res) => {
 
 export const deleteSatuanKerja = async (req, res) => {
     try {
-        await SatuanKerja.destroy({
+        const { id } = req.params
+        const checkSatuanKerja = await SatuanKerja.findOne({
             where: {
-                id: req.params.id
+                id: id
             }
         })
+        if (!checkSatuanKerja) return payload(400, false, "Satuan kerja tidak ditemukan", null, res)
+
+        await SatuanKerja.destroy({
+            where: {
+                id: id
+            }
+        })
+
         return payload(200, true, "Berhasil menghapus data satuan kerja", null, res)
     } catch (error) {
         return payload(500, false, error.message, null, res)
